requirements.txt
runpod==1.7.9
diffusers>=0.33.0
transformers>=4.47.0
accelerate>=0.34.0
huggingface_hub>=0.27.0
sentencepiece
protobuf
opencv-python-headless
imageio[ffmpeg]
Pillow>=10.0.0
numpy
requests